Update variable names in remove user follow sql script

+++ server/APIv1/follows/removeUserFollow.sql
@@ -1,6 +1,6 @@
 DELETE FROM
   followers_users_users
 WHERE
-  target = ${target}
-  AND follower = ${follower}
+  target = ${targetID}
+  AND follower = ${userID}
 RETURNING *;
